Upgrades sbt to 1.3.9, scala to 2.12.11, genjavadoc to 0.16.
sbt, depending on the terminal you are displaying it in, will issue ANSI
colorization escapes. You may need to add the color=false option to your
environment: e.g., in my .bash_aliases I have

export SBT_OPTS=" -Dsbt.color=false "

Had to fix a number of deprecations.
junit.framework.Assert is deprecated. Changed to org.junit.Assert.

Assert.assertEquals for float/double now requires a delta argument. Supplied
1e-15 where needed.

Since scala 2.12.7, shifting of values by quantities expressed as Long is deprectated. Had to insert ".toInt" calls a number of places to silence these.

Changes to main.yml also fix DAFFODIL-2111 which was javadoc-related errors due
to java 11 occurring on the automated build.

DAFFODIL-2313, DAFFODIL-2111

+++ daffodil-lib/src/test/scala/passera/test/UnsignedCheck.scala
@@ -245,7 +245,7 @@ class UnsignedCheck {
 
   @Test def testLshiftLong = {
     assertTrue(
-      forAll { (a: Int, b: Long) => a.toUInt << (b & 0x1f) == (a << (b & 0x1f)).toUInt }
+      forAll { (a: Int, b: Long) => a.toUInt << (b & 0x1f) == (a << (b & 0x1f).toInt).toUInt }
     )
   }
 
@@ -257,7 +257,7 @@ class UnsignedCheck {
 
   @Test def testLshiftULong = {
     assertTrue(
-      forAll { (a: Int, b: Long) => a.toUInt << (b & 0x1f).toULong == (a << (b & 0x1f)).toUInt }
+      forAll { (a: Int, b: Long) => a.toUInt << (b & 0x1f).toULong == (a << (b & 0x1f).toInt).toUInt }
     )
   }
 
@@ -270,7 +270,7 @@ class UnsignedCheck {
 
   @Test def testRshiftLong = {
     assertTrue(
-      forAll { (a: Int, b: Long) => a.toUInt >> (b & 0x1f) == (a >>> (b & 0x1f)).toUInt }
+      forAll { (a: Int, b: Long) => a.toUInt >> (b & 0x1f) == (a >>> (b & 0x1f).toInt).toUInt }
     )
   }
 
@@ -282,7 +282,7 @@ class UnsignedCheck {
 
   @Test def testRshiftULong = {
     assertTrue(
-      forAll { (a: Int, b: Long) => a.toUInt >> (b & 0x1f).toULong == (a >>> (b & 0x1f)).toUInt }
+      forAll { (a: Int, b: Long) => a.toUInt >> (b & 0x1f).toULong == (a >>> (b & 0x1f).toInt).toUInt }
     )
   }
 
@@ -295,7 +295,7 @@ class UnsignedCheck {
 
   @Test def testZrshiftLong = {
     assertTrue(
-      forAll { (a: Int, b: Long) => a.toUInt >>> (b & 0x1f) == (a >>> (b & 0x1f)).toUInt }
+      forAll { (a: Int, b: Long) => a.toUInt >>> (b & 0x1f).toInt == (a >>> (b & 0x1f).toInt).toUInt }
     )
   }
 
@@ -307,7 +307,7 @@ class UnsignedCheck {
 
   @Test def testZrshiftULong = {
     assertTrue(
-      forAll { (a: Int, b: Long) => a.toUInt >>> (b & 0x1f).toULong == (a >>> (b & 0x1f)).toUInt }
+      forAll { (a: Int, b: Long) => a.toUInt >>> (b & 0x1f).toULong == (a >>> (b & 0x1f).toInt).toUInt }
     )
   }
